requirements.txt
[web scrabe]
selenium
bs4
[data parsing]
python-dotenv
bcrypt
snowflake-sqlalchemy
pandas
[data validation]
pydantic

openai

streamlit